Give :bible its docs, its CI wiring and its place in the module lists
The per-extraction bill from the plan, in full:

- bible/AGENT.md and CLAUDE.md, including the thing most likely to trip someone
  up: there are now four Bible-named modules, and this one is the loaded
  translation, not the speech detector (:bible-engine) or the converters that
  produce .spb (:bible-formats).
- test.yml: a paths-filter block WITH its transitive dependency line
  (diagnostics/**, because implementation(projects.diagnostics)), a tests step,
  a coverage-floor step, and :bible:detekt in the unfiltered Detekt step -- the
  module has no baseline and is clean, so it belongs there from day one.
- test-changed.sh: MODULE_DIRS had drifted and was missing :songlibrary and
  :song-chords as well as this module, so a change to any of the three ran no
  suite at all. All three added.
- Root AGENT.md module table and the coverage 'name none' list; README module
  list; FEATURES.md source locations, which now also record why BibleBookNames
  and BibleBookAbbreviations stayed behind.
