+++ bible/CLAUDE.md
@@ -1,0 +1,1 @@
+@AGENT.md
